# api/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home Lab Observer Local API",
    "version": "1.0.0",
    "description": "Read-only, loopback-first API. Non-loopback binding is not part of v1. Log bodies remain local-only and require explicit collection and request opt-in."
  },
  "servers": [{ "url": "http://127.0.0.1:9847" }],
  "security": [{ "localBearer": [] }],
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "getLiveness",
        "summary": "Report process liveness without host details",
        "security": [],
        "responses": {
          "200": {
            "description": "Observer process is live",
            "content": { "application/json": { "schema": { "$ref": "#\/components\/schemas\/Health" } } }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "getReadiness",
        "summary": "Report whether the local API can serve snapshots",
        "security": [],
        "responses": {
          "200": { "description": "Ready", "content": { "application/json": { "schema": { "$ref": "#\/components\/schemas\/Health" } } } },
          "503": { "description": "Not ready", "content": { "application/json": { "schema": { "$ref": "#\/components\/schemas\/Health" } } } }
        }
      }
    },
    "/api/v1/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "summary": "Describe platform support and active privacy policy",
        "x-max-response-bytes": 131072,
        "responses": {
          "200": {
            "description": "Capabilities",
            "content": { "application/json": { "schema": { "$ref": "..\/schemas\/v1\/capabilities-v1.schema.json" } } }
          },
          "401": { "$ref": "#\/components\/responses\/Problem" },
          "500": { "$ref": "#\/components\/responses\/Problem" }
        }
      }
    },
    "/api/v1/snapshots/current": {
      "get": {
        "operationId": "getCurrentSnapshot",
        "summary": "Return one bounded current snapshot",
        "description": "Collection is read-only. include_log_bodies cannot override the local source opt-in, redaction, 2048-byte body limit, or upload prohibition.",
        "x-max-response-bytes": 1048576,
        "parameters": [
          {
            "name": "section", "in": "query", "required": false,
            "description": "Repeat to select sections; omitted means all sections.",
            "style": "form", "explode": true,
            "schema": { "type": "array", "uniqueItems": true, "maxItems": 7, "items": { "enum": ["overview", "filesystems", "processes", "services", "containers", "logs", "observer"] } }
          },
          { "name": "process_limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 } },
          { "name": "container_limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 100 } },
          { "name": "log_limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 0, "maximum": 200, "default": 50 } },
          { "name": "include_log_bodies", "in": "query", "required": false, "schema": { "type": "boolean", "default": false } }
        ],
        "responses": {
          "200": {
            "description": "Current snapshot",
            "content": { "application/json": { "schema": { "$ref": "..\/schemas\/v1\/current-snapshot-v1.schema.json" } } }
          },
          "400": { "$ref": "#\/components\/responses\/Problem" },
          "401": { "$ref": "#\/components\/responses\/Problem" },
          "500": { "$ref": "#\/components\/responses\/Problem" },
          "503": { "$ref": "#\/components\/responses\/Problem" }
        }
      }
    },
    "/api/v1/containers": {
      "get": {
        "operationId": "getContainerInventory",
        "summary": "Return the cached read-only container inventory",
        "description": "Returns the dedicated nullable-metrics container view. It never contacts Docker during an HTTP request and does not change the legacy containers section in the current snapshot, which remains unsupported in v1.",
        "x-max-response-bytes": 1048576,
        "parameters": [
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 100 } }
        ],
        "responses": {
          "200": {
            "description": "Container inventory, including an explicit disabled or unavailable state",
            "content": { "application/json": { "schema": { "$ref": "..\/schemas\/v1\/container-inventory-v1.schema.json" } } }
          },
          "400": { "$ref": "#\/components\/responses\/Problem" },
          "401": { "$ref": "#\/components\/responses\/Problem" },
          "500": { "$ref": "#\/components\/responses\/Problem" }
        }
      }
    },
    "/api/v1/metrics/series": {
      "get": {
        "operationId": "getMetricSeries",
        "summary": "Return bounded local metric series",
        "description": "Returns only code-owned metric identifiers for a fixed time range. Arbitrary metric names, labels, SQL, query languages, and expressions are not accepted. Null values preserve collection gaps and sample_interval_seconds reports the actual stored resolution.",
        "x-max-response-bytes": 1048576,
        "parameters": [
          {
            "name": "range", "in": "query", "required": true,
            "schema": { "type": "string", "enum": ["1h", "6h", "24h", "7d"] }
          },
          {
            "name": "metric", "in": "query", "required": true,
            "description": "Repeat for one to six allowlisted metric identifiers.",
            "style": "form", "explode": true,
            "schema": {
              "type": "array", "minItems": 1, "maxItems": 6, "uniqueItems": true,
              "items": { "type": "string", "enum": ["cpu.utilization.percent", "memory.utilization.percent", "filesystem.aggregate.utilization.percent", "network.receive.bytes_per_second", "network.transmit.bytes_per_second", "process.count"] }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metric series with explicit gaps and actual sample resolution",
            "content": { "application/json": { "schema": { "$ref": "..\/schemas\/v1\/metric-series-v1.schema.json" } } }
          },
          "400": { "$ref": "#\/components\/responses\/Problem" },
          "401": { "$ref": "#\/components\/responses\/Problem" },
          "500": { "$ref": "#\/components\/responses\/Problem" }
        }
      }
    },
    "/api/v1/diagnostics/health": {
      "get": {
        "operationId": "getDiagnosticsHealth",
        "summary": "Report bounded self-diagnostics health and retention counters",
        "description": "Returns code-owned health, limits, storage usage and failure counters only. It never returns diagnostic records, log contents, paths, request data or credentials.",
        "x-max-response-bytes": 32768,
        "responses": {
          "200": {
            "description": "Diagnostics availability, fixed retention limits and bounded counters",
            "content": { "application/json": { "schema": { "$ref": "..\/schemas\/v1\/diagnostics-health-v1.schema.json" } } }
          },
          "401": { "$ref": "#\/components\/responses\/Problem" },
          "500": { "$ref": "#\/components\/responses\/Problem" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "localBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "opaque local token", "description": "Random per-install token. Never accepted in a query parameter." }
    },
    "responses": {
      "Problem": {
        "description": "RFC 9457 problem without secrets or stack details",
        "content": { "application/problem+json": { "schema": { "$ref": "..\/schemas\/v1\/problem-details-v1.schema.json" } } }
      }
    },
    "schemas": {
      "Health": {
        "type": "object", "additionalProperties": false, "required": ["status"],
        "properties": { "status": { "enum": ["UP", "NOT_READY"] } }
      }
    }
  }
}